backend/python/requirements.txt
fastapi==0.118.0
google-cloud-bigquery==3.38.0
google-cloud-pubsub==2.31.1
google-cloud-storage==3.4.0
google-genai==1.41.0
pydantic==2.11.10
pydantic-settings==2.11.0
python-dotenv==1.1.1
python-multipart==0.0.20
uvicorn==0.37.0
